+++ fltk-sys/src/lib.rs
@@ -8,7 +8,6 @@ pub mod dialog;
 pub mod draw;
 pub mod fl;
 pub mod frame;
-pub mod glu;
 pub mod group;
 pub mod image;
 pub mod input;