+++ src/lib.rs
@@ -22,12 +22,12 @@
 //!
 //! ```toml
 //! [dependencies]
-//! fltk = "^0.7"
+//! fltk = "^0.8"
 //! ```
 //! The library offers prebuilt static cfltk and fltk libraries, which can be added using the "fltk-bundled" flag:
 //! ```toml
 //! [dependencies]
-//! fltk = { version = "^0.7", features = ["fltk-bundled"] }
+//! fltk = { version = "^0.8", features = ["fltk-bundled"] }
 //! ```
 //! Since these are pre-built libraries using the Github actions CI, the currently supported operating systems are:
 //! - Windows 10 x64 (msvc and gnu).
@@ -36,11 +36,11 @@
 //! The library is automatically statically linked to your binary. If however you would prefer dynamic linking, you can use the fltk-shared feature:
 //! ```toml
 //! [dependencies.fltk]
-//! version = "^0.7"
+//! version = "^0.8"
 //! features = ["fltk-shared"]
 //! # or
 //! [dependencies]
-//! fltk = { version = "^0.7", features = ["fltk-shared"] }
+//! fltk = { version = "^0.8", features = ["fltk-shared"] }
 //! ```
 //! You can also enable ninja builds for a faster build of the C++ source using the "use-ninja" feature. Or if you have fltk already installed, you can use the fltk-system feature.
 //!
@@ -184,7 +184,7 @@
 //! - system-zlib: Uses the system zlib
 //! - legacy-opengl: Support of Lagacy OpenGL, the crate uses GLVND by default
 //! - fltk-bundled: Support for bundled versions of cfltk and fltk on selected platforms
-//! - no-opengl: Support for systems without OpenGL.
+//! - enable-glwindow: Support for systems without OpenGL.
 //!
 //! ## FAQ
 //!
@@ -197,7 +197,6 @@ pub mod dialog;
 pub mod draw;
 pub mod enums;
 pub mod frame;
-pub mod gl;
 pub mod group;
 pub mod image;
 pub mod input;